+++ TIA/src/tia/Agente.java
@@ -14,9 +14,9 @@ public abstract class Agente {
     protected String nombre;
     protected int edad;
     protected String direccion;
-    protected int salario;
+    protected double salario;
 
-    public Agente(String nombre, int edad, String direccion, int salario) {
+    public Agente(String nombre, int edad, String direccion, double salario) {
         this.nombre = nombre;
         this.edad = edad;
         this.direccion = direccion;
@@ -68,7 +68,7 @@ public abstract class Agente {
     /**
      * @return the salario
      */
-    public int getSalario() {
+    public double getSalario() {
         return salario;
     }
